feat(user-profile): implement new profile, portfolio, and showcase client components
- Added ProfileClient, PortfolioEditClient, ShowcaseClient, ShowcaseDetailClient, and PreviewClient components
- Refactored related page files to use new client components
- Updated layout files across admin and user sections for improved routing and structure
- Added new useAuth hook and auth types
- Implemented middleware for user authentication
- Improved navbar and profile dropdown components
