+++ app/(admin-community)/FAQ-admin-community/layout.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import Footer from "@/components/landing-page/Footer";
+import Footer from "@/components/user/landing-page/Footer";
 import Navbar from "@/components/navbar/AdminCommunityNavbar";
 import { ReactNode } from "react";
