+++ app/(admin-community)/dashboard-admin-community/layout.tsx
@@ -1,6 +1,7 @@
 // dashboard/layout.tsx
-import Footer from "@/components/landing-page/Footer";
+
 import Navbar from "@/components/navbar/AdminCommunityNavbar";
+import Footer from "@/components/user/landing-page/Footer";
 import { ReactNode } from "react";
