+++ app/(admin-community)/home-admin-community/layout.tsx
@@ -1,7 +1,8 @@
 // app/home/layout.tsx
 
-import Footer from "@/components/landing-page/Footer";
+
 import Navbar from "@/components/navbar/AdminCommunityNavbar";
+import Footer from "@/components/user/landing-page/Footer";
 
 
 export default function HomeLayout({ children }: { children: React.ReactNode }) {
