+++ app/(admin-community)/profile-admin-community/layout.tsx
@@ -1,5 +1,5 @@
 // profile/layout.tsx
-import Footer from "@/components/landing-page/Footer";
+import Footer from "@/components/user/landing-page/Footer";
 import Navbar from "@/components/navbar/AdminCommunityNavbar";
 import { ReactNode } from "react";
