+++ app/(admin-community)/showcase-community-admin-community/layout.tsx
@@ -1,8 +1,7 @@
 // layout/ShowcaseLayout.tsx
 "use client";
-
-import Footer from "@/components/landing-page/Footer";
 import Navbar from "@/components/navbar/AdminCommunityNavbar";
+import Footer from "@/components/user/landing-page/Footer";
 import { ReactNode } from "react";
