+++ app/(user)/showcase/layout.tsx
@@ -1,0 +1,3 @@
+export default function ShowcaseLayout({ children }: { children: React.ReactNode }) {
+  return <>{children}</>;
+}
