+++ app/(user)/showcase/page.tsx
@@ -1,39 +1,77 @@
-'use client';
-
-import { useState } from 'react';
+import { cookies } from 'next/headers';
+import { redirect } from 'next/navigation';
 import Footer from '@/components/user/landing-page/Footer';
-import { ShowcaseHeader } from '@/components/user/showcase/ShowcaseHeader';
-import { ShowcaseSearch } from '@/components/user/showcase/ShowcaseSearch';
-import { ShowcaseGrid } from '@/components/user/showcase/ShowcaseGrid';
-import { portfolioItems } from '@/data/portfolioItems'; // Pindahkan data ke file terpisah
+import ShowcaseClient from './ShowcaseClient';
 
-export default function ShowcasePage() {
-  const [searchQuery, setSearchQuery] = useState('');
-  const [category, setCategory] = useState('all');
-
-  const filteredItems = portfolioItems.filter(item => {
-    const matchesSearch =
-      item.title.toLowerCase().includes(searchQuery.toLowerCase()) ||
-      (item.contact?.name?.toLowerCase().includes(searchQuery.toLowerCase()) ?? false);
-    const matchesCategory = category === 'all' || item.category === category;
-    return matchesSearch && matchesCategory;
+async function getPortfolioData(token: string) {
+  const response = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/portofolio`, {
+    headers: {
+      'Authorization': `Bearer ${token}`
+    }
   });
+  
+  if (!response.ok) {
+    throw new Error('Failed to fetch portfolio data');
+  }
+  
+  return response.json();
+}
 
-  return (
-    <div className="min-h-screen flex flex-col">
-      <main className="flex-grow bg-gradient-to-b from-[#001B45] via-[#001233] to-[#051F4C] pt-24 pb-16">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-          <ShowcaseHeader />
-          <ShowcaseSearch
-            searchQuery={searchQuery}
-            onSearchChange={e => setSearchQuery(e.target.value)}
-            category={category}
-            onCategoryChange={setCategory}
-          />
-          <ShowcaseGrid items={filteredItems} />
-        </div>
-      </main>
-      <Footer />
-    </div>
-  );
+export default async function ShowcasePage() {
+  const cookieStore = await cookies();
+  const token = cookieStore.get('token')?.value;
+
+  if (!token) {
+    redirect('/auth/login');
+  }
+
+  try {
+    const portfolioData = await getPortfolioData(token);
+    
+    // Transform the API data to match the expected format
+    const transformedData = portfolioData.map((item: any) => ({
+      id: item.id,
+      name: item.anggota[0]?.user?.nim || '',
+      title: item.nama_projek,
+      image: '/portfolio-1.png', // Using static image as requested
+      category: item.kategori,
+      status: item.status,
+      tags: item.tags.map((tag: any) => tag.nama),
+      date: new Date(item.created_at).toLocaleDateString('en-US', {
+        year: 'numeric',
+        month: 'long',
+        day: 'numeric'
+      }),
+      tahun: item.tahun.toString(),
+      subtitle: `${item.kategori} - ${item.tahun}`,
+      description: item.deskripsi,
+      links: item.detail_project.map((link: any) => ({
+        title: link.judul_link,
+        url: link.link_project
+      })),
+      teamMembers: item.anggota.map((member: any) => ({
+        name: member.user.nim,
+        role: member.role
+      })),
+      contact: {
+        name: item.anggota[0]?.user?.nim || '',
+        id: item.anggota[0]?.user?.id || ''
+      }
+    }));
+
+    return (
+      <div className="min-h-screen flex flex-col">
+        <ShowcaseClient initialData={transformedData} />
+        <Footer />
+      </div>
+    );
+  } catch (error) {
+    console.error('Error fetching portfolio data:', error);
+    return (
+      <div className="min-h-screen flex flex-col">
+        <ShowcaseClient initialData={[]} />
+        <Footer />
+      </div>
+    );
+  }
 }
