+++ components/navbar/ProfileDropdown.tsx
@@ -10,48 +10,66 @@ import {
   DropdownMenuSeparator,
 } from "@/components/ui/dropdown-menu";
 import { Button } from "@/components/ui/button";
-import { useUser } from "@/context/UserContext";
+import { useAuth } from "@/hooks/useAuth";
+import { useState } from "react";
 
 export default function ProfileDropdown() {
   const router = useRouter();
-  const { setUserRole } = useUser();
+  const { logout } = useAuth();
+  const [isLoggingOut, setIsLoggingOut] = useState(false);
 
-  const handleLogout = () => {
-    // Clear user role from context and localStorage
-    setUserRole(null);
-    localStorage.removeItem('userRole');
-    // Redirect to home page
-    router.push('/');
+  const handleLogout = async () => {
+    try {
+      setIsLoggingOut(true);
+      await logout();
+      window.location.href = '/';
+    } catch (error) {
+      console.error('Logout failed:', error);
+      setIsLoggingOut(false);
+    }
   };
 
   return (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
-        <Button 
-          variant="ghost" 
-          size="icon" 
-          className="relative w-8 h-8 rounded-full bg-gray-900 text-white hover:bg-blue-800 transition-colors duration-300 data-[state=open]:bg-blue-800"
-        >
-          <User className="w-5 h-5 text-white" />
-        </Button>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent align="end" className="w-40 bg-blue-900 border-blue-800">
-        <DropdownMenuItem 
-          onClick={() => router.push('/profile')} 
-          className="cursor-pointer text-white focus:bg-blue-950 focus:text-white hover:bg-blue-950 hover:text-white"
-        >
-          <User className="mr-2 h-4 w-4 text-white" />
-          <span>Profile</span>
-        </DropdownMenuItem>
-        <DropdownMenuSeparator className="bg-blue-800" />
-        <DropdownMenuItem 
-          onClick={handleLogout} 
-          className="cursor-pointer text-white focus:bg-blue-950 focus:text-white hover:bg-blue-950 hover:text-white"
-        >
-          <LogOut className="mr-2 h-4 w-4 text-white" />
-          <span>Log Out</span>
-        </DropdownMenuItem>
-      </DropdownMenuContent>
-    </DropdownMenu>
+    <>
+      {/* Loading Overlay */}
+      {/* {isLoggingOut && (
+        <div className="fixed inset-0 z-[9999] flex items-center justify-center bg-black bg-opacity-90">
+          <div className="flex flex-col items-center">
+            <div className="animate-spin rounded-full h-12 w-12 border-t-4 border-b-4 border-blue-500 mb-4"></div>
+            <span className="text-white text-lg font-semibold">Logging out...</span>
+          </div>
+        </div>
+      )} */}
+
+      <DropdownMenu>
+        <DropdownMenuTrigger asChild>
+          <Button 
+            variant="ghost" 
+            size="icon" 
+            className="relative w-8 h-8 rounded-full bg-gray-900 text-white hover:bg-blue-800 transition-colors duration-300 data-[state=open]:bg-blue-800"
+          >
+            <User className="w-5 h-5 text-white" />
+          </Button>
+        </DropdownMenuTrigger>
+        <DropdownMenuContent align="end" className="w-40 bg-blue-900 border-blue-800">
+          <DropdownMenuItem 
+            onClick={() => router.push('/profile')} 
+            className="cursor-pointer text-white focus:bg-blue-950 focus:text-white hover:bg-blue-950 hover:text-white"
+          >
+            <User className="mr-2 h-4 w-4 text-white" />
+            <span>Profile</span>
+          </DropdownMenuItem>
+          <DropdownMenuSeparator className="bg-blue-800" />
+          <DropdownMenuItem 
+            onClick={handleLogout}
+            disabled={isLoggingOut}
+            className="cursor-pointer text-white focus:bg-blue-950 focus:text-white hover:bg-blue-950 hover:text-white disabled:opacity-50 disabled:cursor-not-allowed"
+          >
+            <LogOut className="mr-2 h-4 w-4 text-white" />
+            <span>{isLoggingOut ? 'Logging out...' : 'Log Out'}</span>
+          </DropdownMenuItem>
+        </DropdownMenuContent>
+      </DropdownMenu>
+    </>
   );
 }
