+++ components/user/showcase/types.ts
@@ -1,15 +1,15 @@
 import { LucideIcon } from 'lucide-react';
 
 export interface PortfolioItem {
+  id: string | number;
   name?: string;
-  id: number;
   title: string;
   image: string;
   category: string;
-  status?: string;
+  status: string;
   tags: string[];
   date: string;
-  tahun:string ;
+  tahun: string;
   subtitle: string;
   description: string | string[];
   links?: {
