+++ data/portfolioItems.ts
@@ -6,7 +6,7 @@ export const portfolioItems: PortfolioItem[] = [
     name: 'Krisna Bimantoro',
     title: 'UI/UX Healthy Application',
     image: '/portfolio-1.png',
-    category: 'UI/UX Design',
+    category: 'Rekayasa Perangkat Lunak',
     status: 'Verified',
     tags: ['Mobile Application', 'UI/UX Design', 'Design System'],
     date: 'Released November 21, 2023',
@@ -50,9 +50,9 @@ export const portfolioItems: PortfolioItem[] = [
   {
     id: 2,
     name: 'Heykal',
-    title: 'UI/UX Healthy Application',
+    title: 'Analisis Game',
     image: '/portfolio-1.png',
-    category: 'UI/UX Design',
+    category: 'Game Cerdas',
     status: 'Need Revision',
     tags: ['Mobile Application', 'UI/UX Design', 'Design System'],
     date: 'Released November 21, 2023',
@@ -96,9 +96,9 @@ export const portfolioItems: PortfolioItem[] = [
   {
     id: 3,
     name: 'Haidar',
-    title: 'UI/UX Healthy Application',
+    title: 'Analisis Data Penjualan',
     image: '/portfolio-1.png',
-    category: 'UI/UX Design',
+    category: 'Data Sains',
     status: 'Unverified',
     tags: ['Mobile Application', 'UI/UX Design', 'Design System'],
     date: 'Released November 21, 2023',
@@ -142,9 +142,9 @@ export const portfolioItems: PortfolioItem[] = [
   {
     id: 4,
     name: 'Benny',
-    title: 'UI/UX Healthy Application',
+    title: 'Sistem Keamanan Jaringan',
     image: '/portfolio-1.png',
-    category: 'UI/UX Design',
+    category: 'Keamanan Jaringan',
     status: 'Verified',
     tags: ['Mobile Application', 'UI/UX Design', 'Design System'],
     date: 'Released November 21, 2023',
@@ -188,9 +188,9 @@ export const portfolioItems: PortfolioItem[] = [
   {
     id: 5,
     name: 'Krisna Bimantoro',
-    title: 'UI/UX Healthy Application',
+    title: 'Sistem Keamanan Jaringan',
     image: '/portfolio-1.png',
-    category: 'UI/UX Design',
+    category: 'Keamanan Jaringan',
     status: 'Unverified',
     tags: ['Mobile Application', 'UI/UX Design', 'Design System'],
     date: 'Released November 21, 2023',
@@ -233,9 +233,9 @@ export const portfolioItems: PortfolioItem[] = [
   {
     id: 6,
     name: 'Krisna Bimantoro',
-    title: 'UI/UX Healthy Application',
+    title: 'Sistem Keamanan Jaringan',
     image: '/portfolio-1.png',
-    category: 'UI/UX Design',
+    category: 'Keamanan Jaringan',
     status: 'Need Revision',
     tags: ['Mobile Application', 'UI/UX Design', 'Design System'],
     date: 'Released November 21, 2023',
@@ -278,9 +278,9 @@ export const portfolioItems: PortfolioItem[] = [
   {
     id: 7,
     name: 'Krisna Bimantoro',
-    title: 'UI/UX Healthy Application',
+    title: 'Sistem Keamanan Jaringan',
     image: '/portfolio-1.png',
-    category: 'UI/UX Design',
+    category: 'Keamanan Jaringan',
     status: 'Need Revision',
     tags: ['Mobile Application', 'UI/UX Design', 'Design System'],
     date: 'Released November 21, 2023',
